refactor(entries$): move 'sortEntriesByDateCreated$' type to 'entries$' common module
