+++ src/journal-entry/entries$.ts
@@ -13,3 +13,10 @@ export type createEntry$ = (
     id: string,
     body: string,
 ) => Observable<JournalEntry>;
+
+export type sortEntriesByDateCreated$ = (
+    id: string,
+    order: 1 | -1,
+    page: number,
+    limit: number,
+) => Observable<JournalEntry[]>;
