--- src/journal-entry/sortEntriesByDateCreated.type.ts
@@ -1,9 +1,0 @@
-import {JournalEntry} from "./journal-entry.type";
-import { Observable } from "rxjs";
-
-export type sortEntriesByDateCreated$ = (
-    id: string,
-    order: 1 | -1,
-    page: number,
-    limit: number,
-) => Observable<JournalEntry[]>;
